Added comments. Renamed mgdl-alloc to mgdl-memory. Combined mgdl-cache to mgdl-memory. Added more scripting API functions. Fixed rectangle drawing bug. Changed example to use Handles

+++ source/cross/mgdl-gradient.cpp
@@ -28,7 +28,7 @@ void Gradient_PushColor(struct Gradient* gradient, color32 color, float stop)
 }
 
 
-void Gradient_PushColorArray(struct Gradient* gradient, color32* colors, short amount)
+void Gradient_PushColorArray(struct Gradient* gradient, color32* colors, u8 amount)
 {
     ASSERT_DEBUG(gradient != nullptr);
     if (amount < GRADIENT_SIZE)
